+++ src/ppatosca/Learner.java
@@ -24,13 +24,13 @@ public class Learner {
     public Learner() {
     }
 
-    public Learner(String registrationCode, int lowerTime, int upperTime, int atvref, int senint, int visver, int seqglo, ArrayList<Concept> learningGoals) {
+    public Learner(String registrationCode, float lowerTime, float upperTime, int atvref, int senint, int visver, int seqglo, ArrayList<Concept> learningGoals) {
         this.registrationCode = registrationCode;
         
         // Time is in hours than we must convert to seconds
         
-        this.lowerTime = lowerTime*3600;
-        this.upperTime = upperTime*3600;
+        this.lowerTime = (int)(lowerTime*3600);
+        this.upperTime = (int)(upperTime*3600);
         
         this.learningGoals = learningGoals;
         if (atvref == -11 || atvref == -9) {
